video: fix i-frame-interval can not set to float.
STEPS TO REPRODUCE: (please be specific)
1. APP start a video record , and set the i-frame-interval to float like 0.033, set the frame rate to 240fps
2. Analysis if the i-frame-interval is ok
while the framerate is 240fps, and the i-frame-interval is set to 0.033s.
the EXPECTED RESULTS: there is 1 i-frame in 8 frames.
but the OBSERVED RESULTS: there is 1 i-frame in 240 frames

Bug: 183576655

+++ services/mediametrics/statsd_recorder.cpp
@@ -161,8 +161,8 @@ bool statsd_recorder(const mediametrics::Item *item)
         metrics_proto.set_video_bitrate(videoBitRate);
     }
     // int32 kRecorderVideoIframeInterval = "android.media.mediarecorder.video-iframe-interval";
-    int32_t iFrameInterval = -1;
-    if (item->getInt32("android.media.mediarecorder.video-iframe-interval", &iFrameInterval)) {
+    float iFrameInterval = -1.0;
+    if (item->getFloat("android.media.mediarecorder.video-iframe-interval", &iFrameInterval)) {
         metrics_proto.set_iframe_interval(iFrameInterval);
     }
